refactor(frr): extract BGP policy-chain resolution to sibling file
Move the ordered BGP import/export policy-CHAIN machinery out of
policy_render.go into a cohesive sibling bgp_policy_chain.go (#6424).
This is the self-contained #5277 unit that resolves a neighbor's
effective ordered chain (Junos most-specific-wins), derives the composed
route-map name for a chain of length >= 2, and fails the apply CLOSED on
a collision with an operator route-map name.

Moved whole, unchanged: ReservedChainSuffix; hasNonEmptyPolicy,
filterDefinedPolicies; the global/neighbor export+import chain resolvers;
composedChainName, bgpRouteMapRef; collectBGPRouteMapPolicies,
bgpEffectiveChains, collectBGPComposedChains, bgpComposedChainCollision;
renderComposedBGPChains and equalStringSlice. No signatures, logic, or
render output change; the emitted frr.conf is byte-identical.

Validation: go build ./pkg/frr, gofmt, go vet clean. policy_render.go
drops from 1,398 to 1,113 LOC.
